refactor : moved and cut MarketPlace.php->getOffers() in TradeDB.php->getOffers() and two other functions
