+++ tests/dtu/controllers/Trade/MarketPlace/MarketPlaceTest.php
@@ -63,14 +63,6 @@ class MarketPlaceTest extends TestCase
     $this->assertContains('Location: /user/login', $header);
   }
 
-  public function testGetOffersNoSortNoSearch(): void {
-    $_GET = [];
-    $result = MarketPlace::getOffers();
-    $this->assertIsString($result);
-    $this->assertStringContainsString('<section class="offer-grid">', $result);
-  }
-  //TODO: add more tests for getOffers with different sort and search parameters
-
   public function testDoesNotResolveIncorrectPath(){
     $this->assertFalse(MarketPlace::resolve('/offre/wrongpath', 'GET'));
   }
